PR: For SteamVR projects, show Desktop view before opening URLs
SteamVR has an API call that shows the Steam Desktop in VR, so when a URL is opened in a browser, the player actually can see that. This obviously requires that SteamVR is available to the mod.io plugin. By encapsulating the code so that it's only compiled when the STEAM_VR conditional compilation symbol is set, it doesn't break anything if SteamVR is not available (but obviously only has an effect when that symbol is set).