PR: Make changing PluginSettings from build scripts possible
This is a quick fix for enabling changes to PluginSettings from build scripts. The issue was that Data is a struct only accessible via a static getter; so there was no way to change the values from code, which is a requirement in our project because we have multiple games in the same project and simply use different build scripts for the different games.

This does not fix the issue that we need to duplicate the code to load PluginSettings from the resources. It would be preferable to get the PluginSettings singleton-style instead of using it as just a wrapper for data; so instead of writing PluginSettings.data, it would be PluginSettings.Instance.data. That way, the PluginSettings asset could be loaded and marked easily as dirty for the assets database.