PR: Made keyword replacement more safe
The problem with replacing the keywords on load is that this results in the keywords being persistently replaced when using the new SetGlobalValues(...) workflow for changing the settings during build time.

The new approach handles this by making the serialized members private, and doing the keyword replacement the first time the property getter is called. This way, the persistent member variables are never altered and the final path will never be (accidentally) stored.

While at it, I also replaced the duplicate code with a single method.